Add prefix/suffix options and update to WebExtension
This brings the addon into the future by moving away from the
outdated XUL format at the cost of the button being a different
spot, and having to load a full message just to get the header.

This new approach makes a minor change to the default behavior.
The copied message ID now contains the surrounding "<>" which it
didn't before. However, I tested it with git-send-email and it
still works.

This also adds a new options page with the ability to specifiy
a prefix and suffix to go around the Message ID on copy. This
addresses the two pieces feedback I have recieved. It allows
copying the message IDs as gmail links directly, and it allows
copying the full "Message ID: <>" string.