Add prefix/suffix options and update to WebExtension (#4)
Bring the add-on into the future by moving away from the outdated XUL format.
The messageDisplayAction API was added in Thunderbird 68.3 [1] so raise the minimum version to 68.3.

Also add a new options page with the ability to specify a prefix and suffix to go around the Message ID on copy. This addresses the two pieces feedback I have received. It allows copying the ID with the surrounding "Message ID: <>" string, and it allows copying the message IDs as gmail links. In order to better support that second use case also add a URL encoding option.

The WebExtension approach includes the surrounding brackets, so to maintain backwards compatibility the brackets are stripped on copy. Another new option to include them is added.

Also adds a little "ID" icon for fun.

[1] https://thunderbird-webextensions.readthedocs.io/en/68/messageDisplayAction.html